BREAKING CHANGE: Convert EventStore to interface, add on-memory store
